pyproject.toml
[project]
name = "repo-inspector"
version = "0.1.0"
description = "Script responsavel por validar repositorios"
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.44",
    "matplotlib>=3.10.3",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "redis>=6.1.0",
    "requests>=2.32.3",
]
